[MOD] persist func separated
