+++ src/pages/room/Room.jsx
@@ -5,15 +5,7 @@ import { useParams } from 'react-router-dom';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import SignInForm from '@/components/units/room/SignInForm';
 import { useState } from 'react';
-
-const persist = {
-  get: (key) => {
-    JSON.parse(sessionStorage.getItem(key));
-  },
-  set: (key, value) => {
-    sessionStorage.setItem(key, JSON.stringify(value));
-  }
-};
+import persist from '@/utils/persist';
 
 function Room() {
   const { id } = useParams();
